+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/clock_timezone.php
@@ -1,0 +1,7 @@
+<?php
+
+$container->loadFromExtension('framework', [
+    'clock' => [
+        'default_timezone' => 'Europe/Paris',
+    ],
+]);